+++ internal/alternatives/atomic/limiter_test.go
@@ -1,0 +1,108 @@
+package atomic
+
+import (
+	"sync"
+	"testing"
+	"time"
+
+	"go.uber.org/atomic"
+	"go.uber.org/ratelimit/internal/clock"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestRateLimiter(t *testing.T) {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	defer wg.Wait()
+
+	clock := clock.NewMock()
+	rl := New(100, WithClock(clock), WithoutSlack)
+
+	count := atomic.NewInt32(0)
+
+	// Until we're done...
+	done := make(chan struct{})
+	defer close(done)
+
+	// Create copious counts concurrently.
+	go job(rl, count, done)
+	go job(rl, count, done)
+	go job(rl, count, done)
+	go job(rl, count, done)
+
+	clock.AfterFunc(1*time.Second, func() {
+		assert.InDelta(t, 100, count.Load(), 10, "count within rate limit")
+	})
+
+	clock.AfterFunc(2*time.Second, func() {
+		assert.InDelta(t, 200, count.Load(), 10, "count within rate limit")
+	})
+
+	clock.AfterFunc(3*time.Second, func() {
+		assert.InDelta(t, 300, count.Load(), 10, "count within rate limit")
+		wg.Done()
+	})
+
+	clock.Add(4 * time.Second)
+
+	clock.Add(5 * time.Second)
+}
+
+func TestDelayedRateLimiter(t *testing.T) {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	defer wg.Wait()
+
+	clock := clock.NewMock()
+	slow := New(10, WithClock(clock))
+	fast := New(100, WithClock(clock))
+
+	count := atomic.NewInt32(0)
+
+	// Until we're done...
+	done := make(chan struct{})
+	defer close(done)
+
+	// Run a slow job
+	go func() {
+		for {
+			slow.Take()
+			fast.Take()
+			count.Inc()
+			select {
+			case <-done:
+				return
+			default:
+			}
+		}
+	}()
+
+	// Accumulate slack for 10 seconds,
+	clock.AfterFunc(20*time.Second, func() {
+		// Then start working.
+		go job(fast, count, done)
+		go job(fast, count, done)
+		go job(fast, count, done)
+		go job(fast, count, done)
+	})
+
+	clock.AfterFunc(30*time.Second, func() {
+		assert.InDelta(t, 1200, count.Load(), 10, "count within rate limit")
+		wg.Done()
+	})
+
+	clock.Add(40 * time.Second)
+}
+
+func job(rl *limiter, count *atomic.Int32, done <-chan struct{}) {
+	for {
+		rl.Take()
+		count.Inc()
+		select {
+		case <-done:
+			return
+		default:
+		}
+	}
+}
