+++ ratelimit.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2016 Uber Technologies, Inc.
+// Copyright (c) 2016, 2020 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -23,10 +23,7 @@ package ratelimit // import "go.uber.org/ratelimit"
 import (
 	"time"
 
-	"sync/atomic"
-	"unsafe"
-
-	"go.uber.org/ratelimit/internal/clock"
+	"go.uber.org/ratelimit/internal/alternatives/atomic"
 )
 
 // Note: This file is inspired by:
@@ -48,108 +45,16 @@ type Clock interface {
 	Sleep(time.Duration)
 }
 
-type state struct {
-	last     time.Time
-	sleepFor time.Duration
-}
-
-type limiter struct {
-	state unsafe.Pointer
-	//lint:ignore U1000 Padding is unused but it is crucial to maintain performance
-	// of this rate limiter in case of collocation with other frequently accessed memory.
-	padding [56]byte // cache line size - state pointer size = 64 - 8; created to avoid false sharing.
-
-	perRequest time.Duration
-	maxSlack   time.Duration
-	clock      Clock
-}
-
 // Option configures a Limiter.
-type Option func(l *limiter)
+type Option = atomic.Option
 
-// New returns a Limiter that will limit to the given RPS.
-func New(rate int, opts ...Option) Limiter {
-	l := &limiter{
-		perRequest: time.Second / time.Duration(rate),
-		maxSlack:   -10 * time.Second / time.Duration(rate),
-	}
-	for _, opt := range opts {
-		opt(l)
-	}
-	if l.clock == nil {
-		l.clock = clock.New()
-	}
-	initialState := state{
-		last:     time.Time{},
-		sleepFor: 0,
-	}
-	atomic.StorePointer(&l.state, unsafe.Pointer(&initialState))
-	return l
-}
-
-// WithClock returns an option for ratelimit.New that provides an alternate
-// Clock implementation, typically a mock Clock for testing.
-func WithClock(clock Clock) Option {
-	return func(l *limiter) {
-		l.clock = clock
-	}
-}
-
-// WithoutSlack is an option for ratelimit.New that initializes the limiter
-// without any initial tolerance for bursts of traffic.
-var WithoutSlack Option = withoutSlackOption
-
-func withoutSlackOption(l *limiter) {
-	l.maxSlack = 0
-}
-
-// Take blocks to ensure that the time spent between multiple
-// Take calls is on average time.Second/rate.
-func (t *limiter) Take() time.Time {
-	newState := state{}
-	taken := false
-	for !taken {
-		now := t.clock.Now()
-
-		previousStatePointer := atomic.LoadPointer(&t.state)
-		oldState := (*state)(previousStatePointer)
-
-		newState = state{}
-		newState.last = now
-
-		// If this is our first request, then we allow it.
-		if oldState.last.IsZero() {
-			taken = atomic.CompareAndSwapPointer(&t.state, previousStatePointer, unsafe.Pointer(&newState))
-			continue
-		}
-
-		// sleepFor calculates how much time we should sleep based on
-		// the perRequest budget and how long the last request took.
-		// Since the request may take longer than the budget, this number
-		// can get negative, and is summed across requests.
-		newState.sleepFor += t.perRequest - now.Sub(oldState.last)
-		// We shouldn't allow sleepFor to get too negative, since it would mean that
-		// a service that slowed down a lot for a short period of time would get
-		// a much higher RPS following that.
-		if newState.sleepFor < t.maxSlack {
-			newState.sleepFor = t.maxSlack
-		}
-		if newState.sleepFor > 0 {
-			newState.last = newState.last.Add(newState.sleepFor)
-		}
-		taken = atomic.CompareAndSwapPointer(&t.state, previousStatePointer, unsafe.Pointer(&newState))
-	}
-	t.clock.Sleep(newState.sleepFor)
-	return newState.last
-}
-
-type unlimited struct{}
-
-// NewUnlimited returns a RateLimiter that is not limited.
-func NewUnlimited() Limiter {
-	return unlimited{}
-}
-
-func (unlimited) Take() time.Time {
-	return time.Now()
-}
+// We currently have multiple implementations of the rate-limiter - we plan
+// to do more testing & pick the best one (or give people an option to select
+// appropriate one). In the meantime, we're exposing a single limiter.
+var (
+	// New returns a Limiter that will limit to the given RPS.
+	New = atomic.New
+	// WithClock returns an option for ratelimit.New that provides an alternate
+	// Clock implementation, typically a mock Clock for testing.
+	WithClock = atomic.WithClock
+)
